Changed setting size of the JFrame/Applet thing
Made the JFrame/Applet set itself to the correct size at startup, and
removed the WIDTH = 1024 and HEIGHT = 600 lines since they were no
longer read. Artefacts from copying the code from the village game? Also
moved the size of the maze and the scale it is set to into the Program
class.