Changed saved picture naming system
Added the seed, width and height parameters to the saveScreen method, so
that anyone attempting to recreate the pattern can do so. Old method
left in for compatibility, but you can remove it if you don't need it.
Haven't changed any of the .gif stuff, only the .png saving stuff.